% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fair_assessment.R
\name{as.data.frame.fair_assessment}
\alias{as.data.frame.fair_assessment}
\title{Convert a FAIR assessment to a per-metric data frame.}
\usage{
\method{as.data.frame}{fair_assessment}(x, ...)
}
\arguments{
\item{x}{A \code{fair_assessment} object.}

\item{...}{Ignored.}
}
\value{
A data frame with one row per metric.
}
\description{
Convert a FAIR assessment to a per-metric data frame.
}
