% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{rfair_metric_versions}
\alias{rfair_metric_versions}
\title{List the metric versions bundled with rfair.}
\usage{
rfair_metric_versions()
}
\value{
Character vector of available metric versions (e.g. "0.8").
}
\description{
List the metric versions bundled with rfair.
}
\examples{
rfair_metric_versions()
}
