% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfuji-package.R
\docType{package}
\name{rfuji-package}
\alias{rfuji}
\alias{rfuji-package}
\title{rfuji: Assess the FAIRness of Research Data Objects}
\description{
rfuji is a native R implementation of the F-UJI (FAIRsFAIR Research Data
Object Assessment) metrics. Given a persistent identifier or URL, it resolves
the object, harvests metadata from its landing page and from registries such
as DataCite, and scores the result against the FAIRsFAIR metrics. Unlike the
original rfuji client, this version performs the assessment entirely in R and
does not require a running F-UJI server.
}
\details{
The main entry point is \code{\link[=assess_fair]{assess_fair()}}. See the package vignettes and
\url{https://choxos.github.io/rfuji/} for details.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/choxos/rfuji}
  \item \url{https://choxos.github.io/rfuji/}
  \item Report bugs at \url{https://github.com/choxos/rfuji/issues}
}

}
\author{
\strong{Maintainer}: Ahmad Sofi-Mahmudi \email{a.sofimahmudi@gmail.com} (\href{https://orcid.org/0000-0001-6829-0823}{ORCID})

Authors:
\itemize{
  \item Ahmad Sofi-Mahmudi \email{a.sofimahmudi@gmail.com} (\href{https://orcid.org/0000-0001-6829-0823}{ORCID})
  \item Steffen Neumann \email{sneumann@ipb-halle.de} (Author of the original rfuji F-UJI API client)
}

Other contributors:
\itemize{
  \item PANGAEA (Copyright holder of the F-UJI service whose metrics rfuji reimplements) [copyright holder]
}

}
\keyword{internal}
