% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{rfuji_metric_versions}
\alias{rfuji_metric_versions}
\title{List the metric versions bundled with rfuji.}
\usage{
rfuji_metric_versions()
}
\value{
Character vector of available metric versions (e.g. "0.8").
}
\description{
List the metric versions bundled with rfuji.
}
\examples{
rfuji_metric_versions()
}
